Add orders app

+++ ServiceStation/account/views.py
@@ -10,18 +10,25 @@ from .forms import LoginForm, UserRegistrationForm, UserEditForm, ProfileEditFor
 from django.http import JsonResponse 
 from django.views.decorators.http import require_POST
 from django.views.decorators.csrf import csrf_exempt
+from django.http import HttpResponseRedirect
 
 
 @login_required 
+@csrf_exempt
 def profile(request):   
     profile = Profile.objects.get(id=request.user.id)
     cars = request.user.car.all() 
     if request.method == 'POST':
-        car_form = CarForm(request.POST) 
-        if car_form.is_valid():                    
-            new_car = car_form.save(commit=False)
-            new_car.user_id = request.user.id
-            new_car.save()   
+        car = Car(
+            user_id = request.user.id, 
+            make = request.POST.get('make'),
+            model = request.POST.get('model'),
+            year = request.POST.get('year'),
+            vin = request.POST.get('vin')
+            )
+        car.save()
+        car_json = {'id':car.id, 'make':car.make, 'model':car.model, 'year':car.year, 'vin':car.vin}
+        return JsonResponse({'status':'ok', 'car':car_json})
     else:
         car_form = CarForm()
     return render(request,'account/profile.html',{'section': 'profile', 'profile': profile, 'cars':cars, 'car_form':car_form}) 
@@ -33,11 +40,13 @@ def edit(request):
         profile_form = ProfileEditForm(instance=request.user.profile, data=request.POST, files=request.FILES)        
         if user_form.is_valid() and profile_form.is_valid():            
             user_form.save()            
-            profile_form.save()    
+            profile_form.save()   
+        return HttpResponseRedirect('/account/') 
     else:        
         user_form = UserEditForm(instance=request.user)        
         profile_form = ProfileEditForm(instance=request.user.profile)   
-    return render(request,'account/edit.html', {'user_form': user_form,'profile_form': profile_form}) 
+        return render(request,'account/edit.html', {'user_form': user_form,'profile_form': profile_form}) 
+    
 
 def register(request):    
     if request.method == 'POST':        
@@ -49,8 +58,7 @@ def register(request):
             new_user.save()   
             new_profile = profile_form.save(commit=False)
             new_profile.user = new_user
-            new_profile.save()
-            #Profile.objects.create(user=new_user) 
+            new_profile.save() 
             return render(request, 'account/register_done.html', {'new_user': new_user})    
     else:        
         user_form = UserRegistrationForm() 
@@ -64,7 +72,10 @@ def delete_car(request):
         car_id = request.POST.get('id') 
         car = Car.objects.get(id=car_id)
         if car.user_id == request.user.id:
-            car.delete()
+            if request.user.orders.filter(car_id=car_id).count():
+                return JsonResponse({'status':'error'})
+            else:
+                car.delete()
         else:
             car = None
     except:
